Refactor Claim Component
