+++ src/components/claims/ClaimDetails.tsx
@@ -12,6 +12,9 @@ import { PaperClipIcon } from "@heroicons/react/24/outline";
 import Checklist, { ChecklistItem } from "../common/Checklist";
 
 import { JsonViewer } from "@textea/json-viewer";
+import FinancialInfo from "./FinancialInfo";
+import PatientDetails from "./PatientDetails";
+import MedicalInfo from "./MedicalInfo";
 
 export const Tabss = ({ tabs, activeTab, setActiveTab }: any) => {
   return (
@@ -35,198 +38,9 @@ export const Tabss = ({ tabs, activeTab, setActiveTab }: any) => {
   );
 };
 
-interface RejectApproveHandlers {
+export interface RejectApproveHandlers {
   handleReject: typeof handleReject;
   handleApprove: typeof handleApprove;
-}
-
-export function FinancialInfo({
-  claim,
-  ...props
-}: { claim: ClaimDetail } & RejectApproveHandlers) {
-  const medical_info = claim.medical_info;
-  const financial_info = claim.financial_info;
-  const [approvedAmount, setApprovedAmount] = React.useState(
-    parseFloat(
-      (
-        financial_info.approved_amount ||
-        claim.medical_info.approved_amount ||
-        claim.requested_amount
-      )
-        .toString()
-        .replace("INR ", "")
-    )
-  );
-  const [remarks, setRemarks] = React.useState(financial_info.remarks);
-  const status = financial_info.status;
-
-  return (
-    <div className="px-4 py-5 sm:px-6">
-      <dl className="grid grid-cols-1 gap-x-4 gap-y-8 sm:grid-cols-2">
-        <div className="sm:col-span-1">
-          <dt className="text-sm font-medium text-gray-500">
-            Requested Amount
-          </dt>
-          <dd className="mt-1 text-sm text-gray-900">
-            {claim.requested_amount || "-"}
-          </dd>
-        </div>
-        <div className="sm:col-span-1">
-          <dt className="text-sm font-medium text-gray-500">Approved Amount</dt>
-          <dd className="mt-1 text-sm text-gray-900">
-            {financial_info.approved_amount || "-"}
-          </dd>
-        </div>
-        <div className="sm:col-span-1">
-          <dt className="text-sm font-medium text-gray-500">Status</dt>
-          <dd className="mt-1 text-sm text-gray-900">
-            {<StatusChip status={status} />}
-          </dd>
-        </div>
-        {claim.items && claim.items.length > 0 && (
-          <Table
-            title=""
-            headers={["display", "code", "value"]}
-            data={claim.items.map((item: any) => ({
-              id: item.productOrService.coding[0].code,
-              display: item.productOrService.coding[0].display,
-              code: item.productOrService.coding[0].code,
-              value: `${item.unitPrice.value} ${item.unitPrice.currency}`,
-            }))}
-          />
-        )}
-      </dl>
-    </div>
-  );
-}
-
-export function MedicalInfo({
-  claim,
-  ...props
-}: { claim: ClaimDetail } & RejectApproveHandlers) {
-  const status = claim.medical_info.status;
-
-  return (
-    <div className="px-4 py-5 sm:px-6">
-      <dl>
-        <div className="sm:col-span-1">
-          <dt className="text-sm font-medium text-gray-500">Status</dt>
-          <dd className="mt-1 text-sm text-gray-900">
-            {<StatusChip status={status} />}
-          </dd>
-        </div>
-        {claim.diagnosis && claim.diagnosis.length > 0 && (
-          <Table
-            title=""
-            headers={["display", "code", "text"]}
-            data={claim.diagnosis.map((item) => ({
-              id: item.diagnosisCodeableConcept.coding[0].code,
-              display: item.type[0].coding[0].display,
-              code: item.diagnosisCodeableConcept.coding[0].code,
-              text: item.diagnosisCodeableConcept.text,
-            }))}
-          />
-        )}
-      </dl>
-    </div>
-  );
-}
-
-export function PatientDetails({ claim }: { claim: any }) {
-  const includeFields = [
-    "request_id",
-    "request_no",
-    "provider",
-    "name",
-    "insurance_no",
-    "gender",
-    "status",
-    "address",
-  ];
-
-  const supportingFiles = claim.resources.claim.supportingInfo;
-
-  return (
-    <div className="px-4 py-5 sm:px-6">
-      <dl className="grid grid-cols-1 gap-x-4 gap-y-8 sm:grid-cols-2">
-        {Object.entries(claim)
-          .filter(([name, _]) => includeFields.includes(name))
-          .map(([name, detail]: any) => {
-            return (
-              <div className="sm:col-span-1" key={name}>
-                <dt className="text-sm italic text-gray-500">
-                  {properText(name)}
-                </dt>
-                <dd className="text-sm text-black">
-                  {name === "status" ? (
-                    <StatusChip status={detail} />
-                  ) : name === "address" ? (
-                    <div className="">
-                      {detail?.map((a: any, i: number) => {
-                        return (
-                          <div key={i}>
-                            {a.text},
-                            <br />
-                            {a.city}, {a.state}, {a.postalCode},
-                            <br />
-                            {a.country}
-                          </div>
-                        );
-                      }) || "--"}
-                    </div>
-                  ) : (
-                    detail
-                  )}
-                </dd>
-              </div>
-            );
-          })}
-      </dl>
-      <dl className="mt-8">
-        <dt className="font-semibold mb-4">Supporting Files</dt>
-        <dd>
-          {supportingFiles ? (
-            <ul
-              role="list"
-              className="divide-y divide-gray-200 rounded-md border border-gray-200"
-            >
-              {supportingFiles.map((file: any) => {
-                console.log(file);
-                return (
-                  <li className="flex items-center justify-between py-3 pl-3 pr-4 text-sm">
-                    <div className="flex w-0 flex-1 items-center">
-                      <PaperClipIcon
-                        className="h-5 w-5 flex-shrink-0 text-gray-400"
-                        aria-hidden="true"
-                      />
-                      <a
-                        className="ml-2 w-0 flex-1 truncate hover:text-indigo-500"
-                        href={file.valueAttachment.url}
-                        target="_blank"
-                      >
-                        {file.valueAttachment.title || "File"}
-                      </a>
-                    </div>
-                    <div className="ml-4 flex-shrink-0">
-                      <a
-                        href={file.valueAttachment.url}
-                        download
-                        className="font-medium text-indigo-600 hover:text-indigo-500"
-                      >
-                        Download
-                      </a>
-                    </div>
-                  </li>
-                );
-              })}
-            </ul>
-          ) : (
-            <p className="text-gray-500 text-sm">No supporting files</p>
-          )}
-        </dd>
-      </dl>
-    </div>
-  );
 }
 
 const handleReject = async ({ request_id, type }: any) => {
